requirements.txt
alembic==1.15.1
bcrypt==4.3.0
blinker==1.9.0
cachelib==0.13.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.7.1
Deprecated==1.2.18
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Limiter==3.12
Flask-Migrate==4.1.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.1.0
Mako==1.3.9
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
ordered-set==4.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
Pygments==2.19.1
PyJWT==2.10.1
pytest==8.3.5
pytest-cov==6.0.0
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
rich==13.9.4
SQLAlchemy==2.0.39
typing_extensions==4.12.2
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2
